pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-paper-tracker"
version = "1.0.0"
description = "Daily academic paper tracking system for top journals."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "apscheduler>=3.10.4",
  "fastapi>=0.111.0",
  "feedparser>=6.0.11",
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.3.4",
  "python-dateutil>=2.9.0.post0",
  "python-multipart>=0.0.9",
  "rapidfuzz>=3.9.4",
  "sqlalchemy>=2.0.31",
  "tenacity>=8.5.0",
  "uvicorn[standard]>=0.30.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.7"
]

[project.scripts]
daily-paper = "app.cli:main"

[tool.setuptools]
package-dir = {"" = "apps/api"}

[tool.setuptools.packages.find]
where = ["apps/api"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
pythonpath = [
  "apps/api"
]
testpaths = [
  "apps/api/tests"
]
